Remove Bootstrap class and update _bootstrap.php
