+++ phpunit.xml
@@ -18,7 +18,6 @@
         <whitelist processUncoveredFilesFromWhitelist="true">
             <directory suffix=".php">./system</directory>
             <exclude>
-                <file>./system/Bootstrap.php</file>
                 <file>./system/ComposerScripts.php</file>
                 <file>./system/View/Escaper.php</file>
                 <directory suffix=".php">./system/View/Exception</directory>
