--- tests/_support/MockBootstrap.php
@@ -1,6 +1,0 @@
-<?php namespace CodeIgniter;
-
-class MockBootstrap extends Bootstrap
-{
-
-}
